test(backend): lock Vercel eva-env.sh single-quote escaping
Prevents values with apostrophes from breaking sourced sandbox env files.
